+++ config/settings/base.py
@@ -149,7 +149,7 @@ INSTALLED_APPS = [
     "hexa.countries",
     "hexa.notebooks",
     "hexa.pipelines",
-    "hexa.template_pipelines",
+    "hexa.pipeline_templates",
     "hexa.comments",
     "hexa.tags",
     "hexa.ui",
