+++ hexa/pipeline_templates/apps.py
@@ -2,5 +2,5 @@ from hexa.app import CoreAppConfig
 
 
 class TemplatePipelinesConfig(CoreAppConfig):
-    name = "hexa.template_pipelines"
-    label = "template_pipelines"
+    name = "hexa.pipeline_templates"
+    label = "pipeline_templates"
