+++ hexa/pipeline_templates/graphql/schema.graphql
