+++ hexa/pipeline_templates/permissions.py
