+++ hexa/pipeline_templates/schema/__init__.py
