+++ hexa/pipeline_templates/schema/queries.py
