+++ hexa/pipeline_templates/tests/test_schema/test_templates.py
